refactor: lending to /components
