+++ apps/dapp/src/components/lending/Debt/RepayDialog.tsx
@@ -1,23 +1,24 @@
-import { useState, useCallback } from 'react';
-import Box from '@mui/material/Box';
-import { SsovV3LendingPut__factory } from 'mocks/factories/SsovV3LendingPut__factory';
+import { useCallback, useState } from 'react';
+
 import { BigNumber } from 'ethers';
 
-import EstimatedGasCostButton from 'components/common/EstimatedGasCostButton';
-import { Typography, Input, CustomButton, Dialog } from 'components/UI';
-import ContentRow from 'components/atlantics/InsuredPerps/ManageCard/ManagePosition/ContentRow';
-
+import Box from '@mui/material/Box';
+import useSendTx from 'hooks/useSendTx';
+import { SsovV3LendingPut__factory } from 'mocks/factories/SsovV3LendingPut__factory';
 import { useBoundStore } from 'store';
+
 import { IDebtPosition } from 'store/Vault/lending';
 
-import useSendTx from 'hooks/useSendTx';
+import { CustomButton, Dialog, Input, Typography } from 'components/UI';
+import ContentRow from 'components/atlantics/InsuredPerps/ManageCard/ManagePosition/ContentRow';
+import EstimatedGasCostButton from 'components/common/EstimatedGasCostButton';
 
-import { formatAmount } from 'utils/general';
 import {
   getContractReadableAmount,
   getReadableTime,
   getUserReadableAmount,
 } from 'utils/contracts';
+import { formatAmount } from 'utils/general';
 
 import { ARBITRUM_CHAIN_ID, DECIMALS_TOKEN } from 'constants/index';
